refactoring and first try
refactored the chess rules in to seperate class. game controller in one class.
first implementation of alpha beta pruning engine. can control moves but always choses last one. why??
basic eval rules
